ASD is not starting automatically during BMC boot
Issue:
Security manager randomly fails to the start the ASD.

Security-manager is controlling the ASD service based on
special user enable and Remote debug Jumper.
In order to support WolfPass system, we are reading the
Product ID from EntityManager.

There is very tiny window where EntityManager starts and a
security manager fails to read the product id (As Entity manager is
just start and SubTree Object is not created)
which lead to security-manager fails and unable to recover by itself.
So ASD service is not started automatically.

Fix: Security-manager should wait for InterfaceAdded signal even
when it fails to read the Entity SubTree Object Mapper.

Tested on Whitley:

Check asd service starts when special user and Jumper is set

1. Set the jumper and Set the special user
2. ASD service should run automatically.

Check the security-manager detect the product ID when entity manager
is not started.
1. Stop the entity manager and security-manager and kill asd.
2. Run the secuirty-manager.
3. Security-manager should wait for InterfaceAdded signal.
4. ASD sevice is not started automatically.
5. Start the entity-manager.
6. security-manager should detect the ProductID and start the ASD.

Check the security-manager detect the Product ID when entity- manager
is started already.
1. Stop the security-manager and kill the asd.
2. Run the security-manager.
4. Security-manager should able to read the Product Id from Entity mgr.
5. Start the ASD.

Change-Id: I22ada71b3d8f308a1545314cd7f0913a001a7f83
Signed-off-by: Suryakanth Sekar <suryakanth.sekar@linux.intel.com>